chore: rename lib to src
