+++ examples/every_10_minutes.js
@@ -1,7 +1,7 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('Before job instantiation');
-const job = new CronJob('0 */10 * * * *', function() {
+const job = new CronJob('0 */10 * * * *', function () {
 	const d = new Date();
 	console.log('Every Tenth Minute:', d);
 });
