+++ examples/every_30_minutes_between_9_and_5.js
@@ -1,7 +1,7 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('Before job instantiation');
-const job = new CronJob('0 */30 9-17 * * *', function() {
+const job = new CronJob('0 */30 9-17 * * *', function () {
 	const d = new Date();
 	console.log('Every 30 minutes between 9-17:', d);
 });
