+++ examples/mon_to_fri_at_11_30.js
@@ -1,7 +1,7 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('Before job instantiation');
-const job = new CronJob('00 30 11 * * 1-5', function() {
+const job = new CronJob('00 30 11 * * 1-5', function () {
 	const d = new Date();
 	console.log('onTick:', d);
 });
