+++ examples/multiple_jobs.js
@@ -1,12 +1,12 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('Before job instantiation');
-const job = new CronJob('*/5 * * * * *', function() {
+const job = new CronJob('*/5 * * * * *', function () {
 	const d = new Date();
 	console.log('First:', d);
 });
 
-const job2 = new CronJob('*/8 * * * * *', function() {
+const job2 = new CronJob('*/8 * * * * *', function () {
 	const d = new Date();
 	console.log('Second:', d);
 });
