+++ examples/run_at_specific_date.js
@@ -1,9 +1,9 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('Before job instantiation');
-let date = new Date();
-date.setSeconds(date.getSeconds()+2);
-const job = new CronJob(date, function() {
+const date = new Date();
+date.setSeconds(date.getSeconds() + 2);
+const job = new CronJob(date, function () {
 	const d = new Date();
 	console.log('Specific date:', date, ', onTick at:', d);
 });
