+++ examples/utc_offset_syntax.js
@@ -1,9 +1,9 @@
-const CronJob = require('../lib/cron.js').CronJob;
+const CronJob = require('../src/cron.js').CronJob;
 
 console.log('first');
 const job = new CronJob(
 	'0 0 9 4 * *',
-	function() {
+	function () {
 		console.log('message');
 	},
 	null,
@@ -11,6 +11,6 @@ const job = new CronJob(
 	null,
 	null,
 	null,
-	-360, // UTC-6, represented in minutes
+	-360 // UTC-6, represented in minutes
 );
 console.log('second');
